+++ R/article.R
@@ -175,6 +175,49 @@ frontiers_article <- function(..., keep_tex = TRUE) {
   pdf_document_format("frontiers", keep_tex = keep_tex, ...)
 }
 
+#' @param journal one of \code{"aoas"}, \code{"aap"}, \code{"aop"}, \code{"aos"}, \code{"sts"} for \code{ims_article}
+#' @section \code{ims_article}: Format for creating submissions to the Institute of Mathematical Statistics
+#' \href{https://imstat.org/}{IMS} journals and publications. Adapted from
+#' \url{https://github.com/vtex-soft/texsupport.ims-aoas}.
+#'
+#' The argument \code{journal} accepts the acronym of any of the
+#' \href{https://www.e-publications.org/ims/support/ims-instructions.html}{journals} in IMS:
+#' \itemize{
+#'   \item \code{aap}: The Annals of Applied Probability
+#'   \item \code{aoas}: The Annals of Applied Statistics
+#'   \item \code{aop}: The Annals of Probability
+#'   \item \code{aos}: The Annals of Statistics
+#'   \item \code{sts}: Statistical Science}
+#' @export
+#' @rdname article
+ims_article <- function(journal = c("aoas", "aap", "aop", "aos", "sts"),
+                        keep_tex = TRUE, citation_package = "natbib",
+                        md_extensions = c(
+                          "-autolink_bare_uris" # disables automatic links
+                        ), ...) {
+
+  journal <- match.arg(journal)
+  if (length(journal) > 1) stop("Please choose just one ", dQuote("journal"))
+
+  with_kwsc <- journal %in% c("aap", "aop", "aos") # with keyword_subclass
+
+  args <- c(
+    "journal" = journal,
+    if (with_kwsc) c("with_kwsc" = with_kwsc)
+    )
+
+  # Convert to pandoc arguments
+  pandoc_arg_list <- mapply(rmarkdown::pandoc_variable_arg, names(args), args,
+                            SIMPLIFY = FALSE, USE.NAMES = FALSE)
+  pandoc_arg_list <- unlist(pandoc_arg_list)
+
+  pdf_document_format(
+    "ims", keep_tex = keep_tex, citation_package = citation_package,
+    md_extensions = md_extensions, pandoc_args = pandoc_arg_list,
+    ...
+  )
+}
+
 #' @section \code{jasa_article}: Format for creating submissions to the
 #'   Journal of the Acoustical Society of America. Adapted from
 #'   \url{https://acousticalsociety.org/preparing-latex-manuscripts/}.